TST: additional unit test
